+++ index.js
@@ -1,15 +1,18 @@
 /* eslint-env node, es6 */
 'use strict';
 
+const EmberApp = require('ember-cli/lib/broccoli/ember-app');
+
 module.exports = {
   name: 'ember-cli-pact',
 
   config() {
     return {
       'ember-cli-pact': {
-        mockProvider: 'mirage',
-        providerName: 'provider',
+        providerName: null,
         consumerName: this.project.name(),
+        mockProvider: 'mirage',
+        serviceInjections: ['store'],
         pactsDirectory: 'pacts',
         mode: process.env.PACT_MODE || (process.env.CI ? 'verify' : 'write'),
         pactVersion: 3
@@ -19,7 +22,7 @@ module.exports = {
 
   testemMiddleware(app) {
     const PactMiddleware = require('./lib/pact-middleware');
-    let config = this.project.config()['ember-cli-pact'];
+    let config = this.project.config(EmberApp.env())['ember-cli-pact'];
     let middleware = new PactMiddleware(config);
 
     middleware.attach(app);
