+++ pacts/departments-server.json
@@ -1,0 +1,82 @@
+{
+  "consumer": {
+    "name": "test-consumer"
+  },
+  "interactions": [
+    {
+      "description": "listing departments",
+      "providerStates": [
+        {
+          "name": "a department exists",
+          "params": {
+            "id": "1",
+            "name": "People"
+          }
+        },
+        {
+          "name": "a department exists",
+          "params": {
+            "id": "2",
+            "name": "Admin"
+          }
+        }
+      ],
+      "request": {
+        "headers": {
+          "Accept": "application/vnd.api+json",
+          "X-Requested-With": "XMLHttpRequest"
+        },
+        "method": "GET",
+        "path": "/departments"
+      },
+      "response": {
+        "body": {
+          "data": [
+            {
+              "attributes": {
+                "name": "People"
+              },
+              "id": "1",
+              "relationships": {
+                "members": {
+                  "data": [
+                  ]
+                }
+              },
+              "type": "departments"
+            },
+            {
+              "attributes": {
+                "name": "Admin"
+              },
+              "id": "2",
+              "relationships": {
+                "members": {
+                  "data": [
+                  ]
+                }
+              },
+              "type": "departments"
+            }
+          ]
+        },
+        "headers": {
+          "Content-Type": "application/json"
+        },
+        "matchingRules": {
+          "body": {
+          }
+        },
+        "status": 200
+      }
+    }
+  ],
+  "metadata": {
+    "pact-specification": {
+      "version": "3.0.0"
+    }
+  },
+  "provider": {
+    "name": "departments-server"
+  }
+}
