+++ tests/dummy/config/environment.js
@@ -7,6 +7,12 @@ module.exports = function(environment) {
     environment,
     rootURL: '/',
     locationType: 'auto',
+
+    'ember-cli-pact': {
+      consumerName: 'test-consumer',
+      providerName: 'test-provider'
+    },
+
     EmberENV: {
       FEATURES: {
         // Here you can enable experimental features on an ember canary build
