+++ tests/dummy/mirage/serializers/application.js
@@ -1,5 +1,5 @@
 import { JSONAPISerializer } from 'ember-cli-mirage';
-import PactEnabled from 'ember-cli-pact/mock-provider/mirage/pact-enabled';
+import { PactEnabled } from 'ember-cli-pact/mock-provider/mirage';
 
 export default class ApplicationSerializer extends PactEnabled(JSONAPISerializer) {
