+++ tests/unit/-private/testing/upload-test.js
@@ -4,7 +4,7 @@ import { uploadInteraction } from 'ember-cli-pact/-private/testing/upload';
 module('Unit | upload');
 
 test('a failed upload', function(assert) {
-  return uploadInteraction(null, {})
+  return uploadInteraction(null, { provider: 'x' })
     .then(() => assert.ok(false, 'upload should have failed'))
     .catch((error) => {
       assert.equal(error.message, '[ember-cli-pact] No interaction submitted');
